fix(ci): python-coverage-shards — split oversized provider_recovery test file under 1500-line guard
The task_tag regression test pushed test_provider_recovery_part_001.py to
1518 lines, tripping the first-party 1500-line maintainability guard. Move
the test into a new part_003 so part_001 returns to 1470 lines.
